fixed A*
A* work now for finding the shortest path. The problem was in the
tile.open logic. took out the for loop that add every tile to the open
list, and set the initial tile.open to false.

tile.open was originally(my program) made to see if the tile is in the
opened list without having to traverse the list, and not used to see if
the tile is actually "open" or not.